Add support for serializing `Vec`s in query strings (#30)
This PR adds support for serializing `Vec`s in query strings.

To achieve this we introduce a new `UrlEncodableVec` tuple struct that wraps a `Vec<T>`. We then hand-implement `Serialize` in order to get the proper serialization.

This type is only visible within the crate as its a bit of an implementation detail that we don't want SDK consumers to construct directly. Wherever we use this we'll want to add another wrapper struct that is visible outside of the crate.

Resolves SDK-507.